backend/selector_services/telegram_selector/selector/requirements.txt
psycopg2
python-dotenv
telethon
telebot
lxml
requests
bs4
grpcio 
grpcio-tools 
googleapis-common-protos
asyncio